requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
jax>=0.7.2
jaxlib>=0.4.30
grain
google-cloud-storage>=2.17.0
absl-py
chex
datasets
flax>=0.12.0
optax>=0.2.3
torch>=2.6.0
torchvision>=0.20.1
ftfy
tensorboard>=2.17.0
tensorboardx>=2.6.2.2
tensorboard-plugin-profile>=2.15.2
tokamax
Jinja2
scikit-image
parameterized
Pillow
pylint
pyink
pytest==8.2.2
tensorflow>=2.17.0
tensorflow-datasets>=4.9.6
ruff>=0.1.5,<=0.2
git+https://github.com/Lightricks/LTX-Video
git+https://github.com/zmelumian972/xla@torchax/jittable_module_callable#subdirectory=torchax
opencv-python-headless==4.10.0.84
orbax-checkpoint
tokenizers==0.21.0
huggingface_hub>=0.30.2
transformers==4.48.1
einops==0.8.0
sentencepiece
aqtp
imageio==2.37.0
imageio-ffmpeg==0.6.0
hf_transfer>=0.1.9
qwix@git+https://github.com/google/qwix.git